+++ tests/entrypoints/openai/test_serving_engine.py
@@ -9,7 +9,7 @@ import pytest
 
 from vllm.config import ModelConfig
 from vllm.entrypoints.openai.engine.serving import OpenAIServing
-from vllm.entrypoints.openai.serving_models import OpenAIServingModels
+from vllm.entrypoints.openai.models.serving import OpenAIServingModels
 from vllm.tokenizers.mistral import MistralTokenizer
